+++ src/main/java/profect/goorm1/goormdotcom/config/review/daily/aggregate/steps/aggregate/writer/ReviewDailyAggregateWriter.java
@@ -1,0 +1,7 @@
+package profect.goorm1.goormdotcom.config.review.daily.aggregate.steps.aggregate.writer;
+
+import org.springframework.batch.item.ItemStreamWriter;
+import profect.goorm1.goormdotcom.domain.ReviewDailyAggregate;
+
+public interface ReviewDailyAggregateWriter extends ItemStreamWriter<ReviewDailyAggregate> {
+}
